[JAVA-41972] Moved articles "Arrays.deepEquals", "Intersection Between Two Integer Arrays", "Comparing Arrays in Java", "Performance of System.arraycopy() vs. Arrays.copyOf()", "Calculating the Sum of Two Arrays in Java" from core-java-arrays-operations-advanced to core-java-operations-arrays-advanced-3
