+++ core-java-modules/core-java-arrays-operations-advanced-2/README.md
@@ -1,4 +1,9 @@
-## Relevant Articles
+## Core Java Arrays - Advanced Operations
+
+This module contains articles about advanced operations on arrays in Java. They assume some background knowledge with arrays in Java.
+
+### Relevant Articles:
+
 - [Find the Middle Element of an Array in Java](https://www.baeldung.com/java-array-middle-item)
 - [Find the Equilibrium Indexes of an Array in Java](https://www.baeldung.com/java-equilibrium-index-array)
 - [Moves Zeros to the End of an Array in Java](https://www.baeldung.com/java-array-sort-move-zeros-end)
@@ -7,3 +12,4 @@
 - [Finding the Second Smallest Integer in an Array in Java](https://www.baeldung.com/java-array-second-smallest-integer)
 - [Find the Closest Number to Zero in a Java Array](https://www.baeldung.com/java-array-find-nearest-zero)
 - [Combining Two or More Byte Arrays](https://www.baeldung.com/java-concatenate-byte-arrays)
+- More articles: [[<-- prev]](../core-java-arrays-operations-advanced)[[next -->]](../core-java-arrays-operations-advanced-3)
