+++ core-java-modules/core-java-arrays-operations-advanced-3/README.md
@@ -1,1 +1,12 @@
-## Relevant Articles
+## Core Java Arrays - Advanced Operations
+
+This module contains articles about advanced operations on arrays in Java. They assume some background knowledge with arrays in Java.
+
+### Relevant Articles:
+
+- [Arrays.deepEquals](https://www.baeldung.com/java-arrays-deepequals)
+- [Intersection Between Two Integer Arrays](https://www.baeldung.com/java-array-intersection)
+- [Comparing Arrays in Java](https://www.baeldung.com/java-comparing-arrays)
+- [Performance of System.arraycopy() vs. Arrays.copyOf()](https://www.baeldung.com/java-system-arraycopy-arrays-copyof-performance)
+- [Calculating the Sum of Two Arrays in Java](https://www.baeldung.com/java-sum-arrays-element-wise)
+- More articles: [[<-- prev]](../core-java-arrays-operations-advanced-2)
