+++ core-java-modules/core-java-arrays-operations-advanced-3/src/main/java/com/baeldung/arraycompare/Plane.java
