+++ core-java-modules/core-java-arrays-operations-advanced-3/src/test/java/com/baeldung/arraycompare/LengthsCompareUnitTest.java
@@ -1,8 +1,8 @@
 package com.baeldung.arraycompare;
 
-import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Test;
 
 public class LengthsCompareUnitTest {
