+++ core-java-modules/core-java-arrays-operations-advanced-3/src/test/java/com/baeldung/arraysums/SumArraysUsingForLoopUnitTest.java
@@ -1,8 +1,8 @@
 package com.baeldung.arraysums;
 
-import com.baeldung.arraysums.SumArraysUsingForLoop;
+import static org.junit.Assert.assertArrayEquals;
+
 import org.junit.Test;
-import static org.junit.Assert.*;
 
 public class SumArraysUsingForLoopUnitTest {
