+++ core-java-modules/core-java-arrays-operations-advanced/README.md
@@ -5,11 +5,7 @@ This module contains articles about advanced operations on arrays in Java. They 
 ### Relevant Articles: 
 
 - [How to Copy an Array in Java](https://www.baeldung.com/java-array-copy)
-- [Arrays.deepEquals](https://www.baeldung.com/java-arrays-deepequals)
 - [Find Sum and Average in a Java Array](https://www.baeldung.com/java-array-sum-average)
-- [Intersection Between Two Integer Arrays](https://www.baeldung.com/java-array-intersection)
-- [Comparing Arrays in Java](https://www.baeldung.com/java-comparing-arrays)
 - [Concatenate Two Arrays in Java](https://www.baeldung.com/java-concatenate-arrays)
-- [Performance of System.arraycopy() vs. Arrays.copyOf()](https://www.baeldung.com/java-system-arraycopy-arrays-copyof-performance)
 - [Slicing Arrays in Java](https://www.baeldung.com/java-slicing-arrays)
-- [Calculating the Sum of Two Arrays in Java](https://www.baeldung.com/java-sum-arrays-element-wise)
+- More articles: [[next -->]](../core-java-arrays-operations-advanced-2)
